testando verificao de email, AuthController 80% finalizado

+++ app/Services/AuthService.php
@@ -3,8 +3,12 @@
 namespace App\Services;
 
 use App\Http\Resources\BaseApiResource;
+use App\Mail\TestEmail;
 use App\Models\User;
+use Faker\Provider\Base;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Str;
 
 class AuthService
 {
@@ -27,18 +31,24 @@ class AuthService
     {
         $user = User::create($dados);
         $token = $user->createToken('auth_token')->plainTextToken;
-        return BaseApiResource::success(["user"=>$user, "token"=>$token], meta: ["role"=>$dados["user"]->role, ] );
+        return BaseApiResource::success(["user"=>$user, "token"=>$token], meta: ["role"=>$user->role ] );
 
     }
 
-    public function logout($id){
-        $user =  User::find($id);
-        $user->tokens()->delete();
+    public function logout(){
+        $user = Auth::user();
+       return BaseApiResource::success($user);
+
+//        $user->tokens()->delete();
     }
 
-    public function update($dados, $id)
+    public function update($dados)
     {
-        $user = User::find($id);
+        if($dados["role"] === "admin"){
+            return BaseApiResource::error("Nao autorizado");
+        }
+        $user = Auth::user();
+
         $user->update($dados);
         $user->tokens()->delete();
         $token = $user->createToken('auth_token')->plainTextToken;
@@ -46,9 +56,16 @@ class AuthService
     }
 
     public function delete($id){
-        $user = User::find($id);
+        $user = Auth::user();
         $user->tokens()->delete();
         $user->delete();
+    }
+
+    public function emailValidated($user)
+    {
+        Mail::to('teste@email.com')->send(new TestEmail());
+        $user->email_verification_token = Str::random(60);
+        Mail::to($user->email)->send();
     }
 }
